[Fix] #176 - 댓글 삭제시 적절한 toast 알람으로 수정

+++ client/src/components/feed/comment/tips/ReplyDeleteTips.jsx
@@ -10,34 +10,37 @@ const ReplyDelete = ({ comment, commentsData, setFeedData, feedData }) => {
 
   // DELETE 요청
   const deleteReply = async () => {
-    const configParams = {
-      method: "DELETE",
-      url: `/tip/${tipId}/tipReply/${tipReplyId}/tipComment/${commentId}`,
+    const shouldDelete = window.confirm("정말로 삭제하시겠습니까?");
+    if (shouldDelete) {
+      const configParams = {
+        method: "DELETE",
+        url: `/tip/${tipId}/tipReply/${tipReplyId}/tipComment/${commentId}`,
 
-      headers: {
-        "ngrok-skip-browser-warning": "69420",
-      },
-    };
+        headers: {
+          "ngrok-skip-browser-warning": "69420",
+        },
+      };
 
-    try {
-      const res = await api(configParams);
-      if (res && res.status === 200) {
-        // tipData를 복사해서 업데이트
-        const updatedtipData = JSON.parse(JSON.stringify(feedData));
+      try {
+        const res = await api(configParams);
+        if (res && res.status === 200) {
+          // tipData를 복사해서 업데이트
+          const updatedtipData = JSON.parse(JSON.stringify(feedData));
 
-        updatedtipData.replies.forEach((reply) => {
-          if (reply.tipReplyId === tipReplyId) {
-            reply.comments = reply.comments.filter(
-              (comment) => comment.tipCommentId !== commentId
-            ); // 답글을 삭제
-          }
-        });
-        setFeedData(updatedtipData); // 상태를 업데이트
-        toast.success("답글을 삭제했습니다.");
+          updatedtipData.replies.forEach((reply) => {
+            if (reply.tipReplyId === tipReplyId) {
+              reply.comments = reply.comments.filter(
+                (comment) => comment.tipCommentId !== commentId
+              ); // 답글을 삭제
+            }
+          });
+          setFeedData(updatedtipData); // 상태를 업데이트
+          toast.success("답글을 삭제했습니다.");
+        }
+      } catch (err) {
+        console.error("err comment:", err);
+        toast.error("답글을 삭제할 수 없습니다.");
       }
-    } catch (err) {
-      console.error("err comment:", err);
-      toast.error("답글을 삭제할 수 없습니다.");
     }
   };
 
@@ -47,7 +50,8 @@ const ReplyDelete = ({ comment, commentsData, setFeedData, feedData }) => {
         className="mx-1 hover:font-semibold"
         onClick={() => {
           deleteReply();
-        }}>
+        }}
+      >
         삭제하기
       </button>
     </div>
